requirements_chairdpo_train.txt
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
bitsandbytes==0.45.0
blis==1.2.0
catalogue==2.0.10
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
deepspeed==0.12.6
dill==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
einops-exts==0.0.4
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
evaluate==0.4.3
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.13.1
flash-attn==2.6.2
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
hf_transfer==0.1.6
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.27.0
idna==3.4
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.77.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
narwhals==1.18.4
networkx==3.2.1
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
orjson==3.10.12
packaging==24.2
pandas==2.2.3
peft==0.12.0
pillow==10.2.0
platformdirs==4.3.6
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pynvml==12.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.3
s2wrapper @ git+https://github.com/bfshi/scaling_on_scales.git@3bbee376f98fcd0e670ef6814416e35d5454e803
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.41.3
svgwrite==1.4.3
sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.67.1
transformers==4.43.1
triton==2.3.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.13
uvicorn==0.34.0
wandb==0.19.1
wasabi==1.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
websockets==11.0.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
